gs_effect_main created, app_main cleaned up
