+++ components/gs_effects/CMakeLists.txt
@@ -1,3 +1,3 @@
 idf_component_register(SRC_DIRS "src"
                        INCLUDE_DIRS "include"
-                       PRIV_REQUIRES jsmn log)
+                       PRIV_REQUIRES jsmn log driver ch_buttons gs_i2s audio_board audio_hal audio_sal esp_peripherals)
