+++ components/gs_i2s/CMakeLists.txt
@@ -1,3 +1,3 @@
 idf_component_register(SRC_DIRS "src"
                        INCLUDE_DIRS "include"
-                       PRIV_REQUIRES log driver audio_board esp_peripherals)
+                       PRIV_REQUIRES log driver audio_board audio_hal audio_sal esp_peripherals)
